requirements.txt
parsedatetime==2.1
gevent==1.1.1
greenlet==0.4.9
beautifulsoup4==4.4.1
lxml==3.6.0
requests==2.10.0
redis==2.10.5
boto3==1.5.9
oauth2client==4.1.3
mysql-connector==2.2.9